Rename wasi.ts -> debug.ts
